广告主和媒体查询接口

+++ core/src/main/java/com/madhouse/platform/premiummad/model/MediaModel.java
@@ -1,27 +1,27 @@
 package com.madhouse.platform.premiummad.model;
 
 public class MediaModel {
-	private long id;
+	private Integer id;
 	private String name;
-	private int category;
+	private Byte category;
 	// app, site
-	private int type;
-	private int advertiserAuditMode;
-	private int materialAuditMode;
+	private Byte type;
+	private Byte advertiserAuditMode;
+	private Byte materialAuditMode;
 
-	public int getType() {
+	public Byte getType() {
 		return type;
 	}
 
-	public void setType(int type) {
+	public void setType(Byte type) {
 		this.type = type;
 	}
 
-	public long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
 
@@ -33,27 +33,27 @@ public class MediaModel {
 		this.name = name;
 	}
 
-	public int getCategory() {
+	public Byte getCategory() {
 		return category;
 	}
 
-	public void setCategory(int category) {
+	public void setCategory(Byte category) {
 		this.category = category;
 	}
 
-	public int getAdvertiserAuditMode() {
+	public Byte getAdvertiserAuditMode() {
 		return advertiserAuditMode;
 	}
 
-	public void setAdvertiserAuditMode(int advertiserAuditMode) {
+	public void setAdvertiserAuditMode(Byte advertiserAuditMode) {
 		this.advertiserAuditMode = advertiserAuditMode;
 	}
 
-	public int getMaterialAuditMode() {
+	public Byte getMaterialAuditMode() {
 		return materialAuditMode;
 	}
 
-	public void setMaterialAuditMode(int materialAuditMode) {
+	public void setMaterialAuditMode(Byte materialAuditMode) {
 		this.materialAuditMode = materialAuditMode;
 	}
 }
